edge-jobs/psychometrics/requirements.txt
psycopg2-binary>=2.9.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
